requirements.txt
pyzmq
mypy
matplotlib
numpy
scipy